refactor: simplify build and packaging of pagination
