+++ packages/components/pagination/bs-page-link.js
@@ -1,6 +1,6 @@
 
 import { LitElement, html } from 'lit-element';
-import { BsPageLinkCss } from './bs-page-link.css.js';
+import { BsPageLinkCss } from './css/bs-page-link.css.js';
 import { BsContentRebootCss } from '@lit-element-bootstrap/content';
 
 export class BsPageLink extends LitElement {
