+++ packages/components/pagination/css/bs-pagination.css.js
