+++ packages/components/pagination/css/index.js
@@ -1,6 +1,3 @@
 export { BsPageItemCss } from './bs-page-item.css.js';
-export { BsPageItem } from './bs-page-item.js';
 export { BsPageLinkCss } from './bs-page-link.css.js';
-export { BsPageLink } from './bs-page-link.js';
 export { BsPaginationCss } from './bs-pagination.css.js';
-export { BsPagination } from './bs-pagination.js';
