+++ packages/components/pagination/index.js
@@ -1,0 +1,4 @@
+export * from './css'
+export { BsPageItem } from './bs-page-item.js';
+export { BsPageLink } from './bs-page-link.js';
+export { BsPagination } from './bs-pagination.js';
